refactor(Participant): Dealer와 Player의 공통 기능 분리
- Participant 부모 클래스 생성
