+++ src/main/java/blackjack/domain/Dealer.java
@@ -1,34 +1,8 @@
 package blackjack.domain;
 
-public class Dealer {
-
-    private static final int BLACK_JACK_SCORE = 21;
-    private static final int ACE_ALTER_VALUE = 10;
-
-    private final Cards cards;
+public class Dealer extends Participant{
 
     public Dealer() {
-        this.cards = new Cards();
+        super();
     }
-
-    public void addCard(Card card) {
-        cards.add(card);
-    }
-
-    public int calculateScore() {
-        int sum = cards.sum();
-        int aceCount = cards.getAceCount();
-
-        while (sum > BLACK_JACK_SCORE && aceCount > 0) {
-            sum -= ACE_ALTER_VALUE;
-            aceCount -= 1;
-        }
-        return sum;
-    }
-
-    public Cards getCards() {
-        return cards;
-    }
-
-
 }
